Content Updates and Auto Export Refactors
