+++ ppr-ui/package.json
@@ -21,7 +21,7 @@
     "@bcrs-shared-components/input-field-date-picker": "^1.0.0",
     "@lemoncode/fonk": "^1.5.1",
     "@lemoncode/fonk-range-number-validator": "^1.1.0",
-    "@nuxt/eslint": "^0.3.13",
+    "@nuxt/eslint": "1.3.0",
     "@nuxtjs/dotenv": "^1.4.2",
     "@pinia/nuxt": "^0.5.1",
     "@sentry/vue": "^7.50.0",
@@ -68,20 +68,20 @@
     "@pinia/testing": "^0.1.3",
     "@rollup/plugin-commonjs": "^25.0.7",
     "@types/node": "^22.10.1",
-    "@typescript-eslint/eslint-plugin": "^6.8.0",
-    "@typescript-eslint/parser": "^6.8.0",
+    "@typescript-eslint/eslint-plugin": "^8.29.0",
+    "@typescript-eslint/parser": "^8.29.0",
     "@vue/compiler-sfc": "^3.3.4",
-    "@vue/eslint-config-standard": "^8.0.1",
-    "@vue/eslint-config-typescript": "^12.0.0",
+    "@vue/eslint-config-standard": "^9.0.1",
+    "@vue/eslint-config-typescript": "^14.5.0",
     "@vue/test-utils": "^2.4.6",
     "axe-core": "^4.8.1",
     "css-loader": "^7.1.2",
-    "eslint": "^8.51.0",
-    "eslint-plugin-import": "^2.28.1",
+    "eslint": "9.23.0",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-vue": "^9.17.0",
-    "eslint-plugin-vuetify": "^2.0.5",
+    "eslint-plugin-promise": "^7.2.1",
+    "eslint-plugin-vue": "^10.0.0",
+    "eslint-plugin-vuetify": "^2.5.2",
     "happy-dom": "^14.12.0",
     "jsdom": "^22.1.0",
     "maska": "^2.1.10",
