--- ppr-ui/src/components/collateral/index.ts
@@ -1,3 +1,0 @@
-export * from './general'
-export * from './vehicle'
-export { default as Collateral } from './Collateral.vue'
