+++ ppr-ui/tests/unit/ReviewConfirm.spec.ts
@@ -16,7 +16,7 @@ import { FeeSummaryTypes } from '@/composables/fees/enums'
 import { RegistrationLengthTrustSummary } from '@/components/registration'
 import SecuritiesActNoticesPanels from '@/components/registration/securities-act-notices/SecuritiesActNoticesPanels.vue'
 import { Parties } from '@/components/parties'
-import { Collateral } from '@/components/collateral'
+import Collateral from '@/components/collateral/Collateral.vue'
 import type { LengthTrustIF } from '@/interfaces'
 import { RegistrationTypes } from '@/resources'
 import { useStore } from '@/store/store'
