+++ ppr-ui/vitest.config.ts
@@ -1,5 +1,4 @@
 import { defineVitestConfig } from '@nuxt/test-utils/config'
-import { fileURLToPath } from 'node:url'
 
 export default defineVitestConfig({
   test: {
@@ -10,19 +9,14 @@ export default defineVitestConfig({
       }
     },
     globals: true,
-    testTimeout: 60000,
-    silent: false,
+    testTimeout: 100000,
     setupFiles: '../tests/setup.ts',
     onConsoleLog (log) {
       if (log.includes('warn')) return false // Filter out Vue warnings while preserving errors and logs.
-      if (log.includes('error')) return false // Filter out failed network requests vs mocking them all.
     },
     environment: 'nuxt',
     environmentOptions: {
       nuxt: {
-        rootDir: fileURLToPath(new URL('./', import.meta.url)),
-        domEnvironment:
-          (process.env.VITEST_DOM_ENV as 'happy-dom' | 'jsdom') ?? 'happy-dom',
         overrides: {
           runtimeConfig: {
             public: {
